+++ CHANGELOG.rst
@@ -11,7 +11,10 @@ Version 0.1.0.9000
     - Restructure environment variables, remove :code:`HOST` and :code:`PORT` from core socket classes.
     - Move log file handling to the interface, create seperate log file for each test and upload as artifacts in github action package check.
     - Split :code:`interface.py` into a :code:`Server` context class and a :code:`Client` context class.
-    - (To do!) Detailed usage section in README.
+    - Added :code:`__repr__` methods to :code:`Client`, :code:`Server` and :code:`Message` classes.
+    - Define :code:`__slots__` for all classes.
+    - Use :code:`logging` package from the standard library to perform logging tasks. Split logs at level INFO into STDOUT and STDERR on the stream handler for the console.
+    - Detailed usage section in README.
 - Bugfixes:
     - (To do!) Update docstrings.
     - (To do!) Render module index and docstrings on readthedocs.